% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Spin up a Shiny app}
\usage{
run_app(data, model, .xai_folder = NULL)
}
\arguments{
\item{data}{dataset to be given}

\item{model}{model function to be given}
}
\description{
This function calls the shiny app contained within the \code{inst} folder
}
